+++ build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
 }
 
-group 'com.fibremint.blockchain'
+group 'com.fibremint.blockchain.server'
 version '1.0-SNAPSHOT'
 
 sourceCompatibility = 1.8
@@ -16,6 +16,7 @@ dependencies {
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.5'
     compile group: 'org.slf4j', name: 'slf4j-simple', version: '1.7.25'
     compile group: 'org.bouncycastle', name: 'bcprov-jdk15on', version: '1.59'
+    compile group: 'org.bouncycastle', name: 'bcpkix-jdk15on', version: '1.59'
 }
 
 jar {
