--- src/main/java/com/fibremint/blockchain/BlockchainMiner.java
@@ -1,18 +1,0 @@
-package com.fibremint.blockchain;
-
-import com.fibremint.blockchain.blockchain.Wallet;
-
-import javax.swing.*;
-
-public class BlockchainMiner extends JFrame implements Runnable {
-    private Wallet wallet;
-
-    @Override
-    public void run() {
-
-    }
-
-    public static void main(String[] args) {
-
-    }
-}
