--- src/main/java/com/fibremint/blockchain/blockchain/TransactionInput.java
@@ -1,12 +1,0 @@
-package com.fibremint.blockchain.blockchain;
-
-import java.io.Serializable;
-
-public class TransactionInput implements Serializable {
-    public String transactionOutputHash;
-    public TransactionOutput UTXO;
-
-    public TransactionInput(String transactionOutputHash) {
-        this.transactionOutputHash = transactionOutputHash;
-    }
-}
