--- src/main/java/com/fibremint/blockchain/message/model/MessageBalance.java
@@ -1,4 +1,0 @@
-package com.fibremint.blockchain.message.model;
-
-public class MessageBalance {
-}
