--- src/main/java/com/fibremint/blockchain/message/model/MessageBase.java
@@ -1,11 +1,0 @@
-package com.fibremint.blockchain.message.model;
-
-public abstract class MessageBase {
-    public MessageType type;
-
-    public MessageBase(MessageType type) {
-        this.type = type;
-    }
-
-
-}
