--- src/main/java/com/fibremint/blockchain/message/model/MessageMineBlock.java
@@ -1,8 +1,0 @@
-package com.fibremint.blockchain.message.model;
-
-public class MessageMineBlock extends MessageBase {
-    public MessageMineBlock() {
-        super(MessageType.mineBlock);
-
-    }
-}
