--- src/main/java/com/fibremint/blockchain/message/model/MessageResult.java
@@ -1,26 +1,0 @@
-package com.fibremint.blockchain.message.model;
-
-import com.google.gson.annotations.SerializedName;
-
-public class MessageResult extends MessageBase {
-    private Type result;
-    private String description;
-
-    public MessageResult(Type result) {
-        super(MessageType.result);
-        this.result = result;
-    }
-
-    public MessageResult(Type result, String description) {
-        super(MessageType.result);
-        this.result = result;
-        this.description = description;
-    }
-
-    public enum Type {
-        @SerializedName("accepted") accepted,
-        @SerializedName("denied") denied,
-        @SerializedName("error") error
-    }
-
-}
