+++ src/main/java/com/fibremint/blockchain/server/blockchain/TransactionInput.java
@@ -1,0 +1,12 @@
+package com.fibremint.blockchain.server.blockchain;
+
+import java.io.Serializable;
+
+public class TransactionInput implements Serializable {
+    public String transactionOutputHash;
+    public TransactionOutput UTXO;
+
+    public TransactionInput(String transactionOutputHash) {
+        this.transactionOutputHash = transactionOutputHash;
+    }
+}
