+++ src/main/java/com/fibremint/blockchain/server/net/MessageSenderRunnable.java
@@ -1,0 +1,40 @@
+package com.fibremint.blockchain.server.net;
+
+import com.fibremint.blockchain.server.net.ServerInfo;
+
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.net.InetSocketAddress;
+import java.net.Socket;
+
+public class MessageSenderRunnable implements Runnable{
+
+    private ServerInfo destServer;
+    private String message;
+
+    public MessageSenderRunnable(ServerInfo destServer, String message) {
+        this.destServer = destServer;
+        this.message = message;
+    }
+
+    @Override
+    public void run() {
+        try {
+            // create socket with a timeout of 2 seconds
+            Socket s = new Socket();
+            s.connect(new InetSocketAddress(this.destServer.getHost(), this.destServer.getPort()), 10000);
+            PrintWriter pw =  new PrintWriter(s.getOutputStream(), true);
+            
+            // send the message forward
+            System.out.println(message);
+        	pw.println(message);
+        	pw.flush();
+
+            // close printWriter and socket
+            pw.close();
+            s.close();
+            
+        } catch (IOException e) {
+        }
+    }
+}
