+++ src/main/java/com/fibremint/blockchain/server/net/message/MessageBalance.java
@@ -1,0 +1,4 @@
+package com.fibremint.blockchain.server.net.message;
+
+public class MessageBalance {
+}
