+++ src/main/java/com/fibremint/blockchain/server/net/message/MessageBase.java
@@ -1,0 +1,11 @@
+package com.fibremint.blockchain.server.net.message;
+
+public abstract class MessageBase {
+    public MessageType type;
+
+    public MessageBase(MessageType type) {
+        this.type = type;
+    }
+
+
+}
