+++ src/main/java/com/fibremint/blockchain/server/net/message/MessageCatchUp.java
@@ -1,0 +1,11 @@
+package com.fibremint.blockchain.server.net.message;
+
+public class MessageCatchUp extends MessageBase{
+    public int blockIndex;
+
+    public MessageCatchUp(int blockIndex) {
+        super(MessageType.catchUp);
+        this.blockIndex = blockIndex;
+
+    }
+}
