+++ src/main/java/com/fibremint/blockchain/server/net/message/MessageHeartbeat.java
@@ -1,0 +1,21 @@
+package com.fibremint.blockchain.server.net.message;
+
+public class MessageHeartbeat extends MessageBase {
+    public int localPort;
+    public int sequenceNumber;
+
+    public MessageHeartbeat(int localPort, int sequenceNumber) {
+        super(MessageType.heartbeat);
+        this.localPort = localPort;
+        this.sequenceNumber = sequenceNumber;
+
+    }
+
+    public int getLocalPort() {
+        return localPort;
+    }
+
+    public int getSequenceNumber() {
+        return sequenceNumber;
+    }
+}
