+++ src/main/java/com/fibremint/blockchain/server/net/message/MessageMineBlock.java
@@ -1,0 +1,7 @@
+package com.fibremint.blockchain.server.net.message;
+
+public class MessageMineBlock extends MessageBase {
+    public MessageMineBlock() {
+        super(MessageType.mineBlock);
+    }
+}
