+++ src/main/java/com/fibremint/blockchain/server/net/message/MessageType.java
@@ -1,0 +1,15 @@
+package com.fibremint.blockchain.server.net.message;
+
+import com.google.gson.annotations.SerializedName;
+
+public enum MessageType {
+    @SerializedName("catchUp") catchUp,
+    @SerializedName("heartbeat") heartbeat,
+    @SerializedName("latestBlock") latestBlock,
+    @SerializedName("mineBlock") mineBlock,
+    @SerializedName("properties") properties,
+    @SerializedName("result") result,
+    @SerializedName("sendFund") sendFund,
+    @SerializedName("serverInQuestion") serverInQuestion,
+    @SerializedName("transaction") transaction
+}
